refactor: split type checker driver
